+++ source/vstd/std_specs/mod.rs
@@ -7,6 +7,7 @@ pub mod borrow;
 pub mod clone;
 pub mod cmp;
 pub mod control_flow;
+pub mod convert;
 pub mod core;
 pub mod ops;
